requirements.txt
python-dotenv
alive-progress
pandas
fsspec
tiktoken
jsonlines
leptonai

# For analysis
datasets
jupyterlab
matplotlib
scikit-learn
choix
seaborn